requirements.txt
pillow==10.4.0
psutil==6.0.0
pyusb==1.2.1
PyYAML==6.0.1
setuptools==71.1.0
